Add multi-version upload and progress to TUI
Introduces multi-version upload workflow and progress bar support in the TUI (cmd/tui_main.go). Adds new upload step state machine, version input fields, and per-version progress tracking. Refactors upload logic to support multiple model versions and improves user interaction for file selection and confirmation. Also adds supporting types and functions for progress reporting. Updates related components and fixes input focus and navigation for the new workflow.

+++ lib/utils.go
@@ -4,9 +4,9 @@ import (
 	"time"
 )
 
-func Throttle(fn func(x int), wait time.Duration) func(x int) {
+func Throttle(fn func(x int64), wait time.Duration) func(x int64) {
 	lastTime := time.Now()
-	return func(x int) {
+	return func(x int64) {
 		now := time.Now()
 		if now.Sub(lastTime) >= wait {
 			fn(x)
